requirements.txt
# Web Framework
Flask==3.1.1
Quart==0.20.0
quart-schema==0.22.0
Hypercorn==0.17.3
Werkzeug==3.1.3

# HTTP/WebSocket Support
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
wsproto==1.2.0
priority==2.0.0

# Template & Security
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0

# Database - PostgreSQL Async
asyncpg==0.29.0
SQLAlchemy[asyncio]==2.0.41
sqlalchemy-utils==0.41.1
alembic==1.16.1
greenlet==3.0.3

# Authentication
PyJWT==2.9.0

# MQTT Client
paho-mqtt==1.6.1

# Data Validation & Serialization
pydantic==2.11.5
pydantic-core==2.33.2
annotated-types==0.7.0
pyhumps==3.8.0

# Utilities
aiofiles==24.1.0
click==8.2.1
colorama==0.4.6
python-dotenv==1.1.0
blinker==1.9.0
typing-inspection==0.4.1
typing-extensions==4.13.2